Improved Bounding Box Export Functionality
